+++ cmd/rabtap/cmd_subscribe_test.go
@@ -32,12 +32,12 @@ func TestCmdSubFailsEarlyWhenBrokerIsNotAvailable(t *testing.T) {
 	go func() {
 		// we expect cmdSubscribe to return
 		cmdSubscribe(ctx, CmdSubscribeArg{
-			amqpURL:            amqpURL,
-			queue:              "queue",
-			tlsConfig:          &tls.Config{},
-			messageReceiveFunc: func(rabtap.TapMessage) error { return nil },
-			termPred:           &constantPred{false},
-			timeout:            time.Second * 10,
+			amqpURL:     amqpURL,
+			queue:       "queue",
+			tlsConfig:   &tls.Config{},
+			messageSink: func(rabtap.TapMessage) error { return nil },
+			termPred:    &constantPred{false},
+			timeout:     time.Second * 10,
 		})
 		done <- true
 	}()
@@ -78,13 +78,13 @@ func TestCmdSub(t *testing.T) {
 
 	// subscribe to testQueue
 	go cmdSubscribe(ctx, CmdSubscribeArg{
-		amqpURL:            amqpURL,
-		queue:              testQueue,
-		tlsConfig:          tlsConfig,
-		messageReceiveFunc: receiveFunc,
-		filterPred:         constantPred{true},
-		termPred:           constantPred{false},
-		timeout:            time.Second * 10,
+		amqpURL:     amqpURL,
+		queue:       testQueue,
+		tlsConfig:   tlsConfig,
+		messageSink: receiveFunc,
+		filterPred:  constantPred{true},
+		termPred:    constantPred{false},
+		timeout:     time.Second * 10,
 	})
 
 	time.Sleep(time.Second * 1)
@@ -100,7 +100,7 @@ func TestCmdSub(t *testing.T) {
 			routingKey: &testKey,
 			headers:    rabtap.KeyValueMap{},
 			tlsConfig:  tlsConfig,
-			providerFunc: func() (RabtapPersistentMessage, error) {
+			source: func() (RabtapPersistentMessage, error) {
 				// provide exactly one message
 				if messageCount > 0 {
 					return RabtapPersistentMessage{}, io.EOF