+++ cmd/rabtap/main.go
@@ -61,7 +61,7 @@ func getTLSConfig(insecureTLS bool, certFile string, keyFile string, caFile stri
 	}
 
 	if caFile != "" {
-		caCert, err := ioutil.ReadFile(caFile)
+		caCert, err := os.ReadFile(caFile)
 		failOnError(err, "invalid tls ca file", os.Exit)
 		caCertPool := x509.NewCertPool()
 		caCertPool.AppendCertsFromPEM(caCert)
@@ -90,12 +90,12 @@ func startCmdInfo(ctx context.Context, args CommandLineArgs, titleURL *url.URL) 
 			out: NewColorableWriter(os.Stdout)})
 }
 
-// createMessageReaderForPublish returns a MessageReaderFunc that reads
+// createMessageReaderForPublish returns a message source that reads
 // messages from the given source in the specified format. The source can
 // be either empty (=stdin), a filename or a directory name
-func createMessageReaderForPublishFunc(source *string, format string) (MessageProviderFunc, error) {
+func newPublishMessageSource(source *string, format string) (MessageSource, error) {
 	if source == nil {
-		return CreateMessageReaderFunc(format, os.Stdin)
+		return NewReaderMessageSource(format, os.Stdin)
 	}
 
 	fi, err := os.Stat(*source)
@@ -109,7 +109,7 @@ func createMessageReaderForPublishFunc(source *string, format string) (MessagePr
 			return nil, err
 		}
 		// TODO close file
-		return CreateMessageReaderFunc(format, file)
+		return NewReaderMessageSource(format, file)
 	} else {
 
 		metadataFiles, err := LoadMetadataFilesFromDir(*source, ioutil.ReadDir, NewRabtapFileInfoPredicate())
@@ -122,7 +122,7 @@ func createMessageReaderForPublishFunc(source *string, format string) (MessagePr
 				metadataFiles[j].metadata.XRabtapReceivedTimestamp)
 		})
 
-		return CreateMessageFromDirReaderFunc(format, metadataFiles)
+		return NewReadFilesFromDirMessageSource(format, metadataFiles)
 	}
 }
 
@@ -130,35 +130,35 @@ func startCmdPublish(ctx context.Context, args CommandLineArgs) {
 	if args.Format == "raw" && args.PubExchange == nil && args.PubRoutingKey == nil {
 		fmt.Fprint(os.Stderr, "Warning: using raw message format but neither exchange or routing key are set.\n")
 	}
-	provider, err := createMessageReaderForPublishFunc(args.Source, args.Format)
-	provider = NewTransformingMessageProvider(provider,
+	source, err := newPublishMessageSource(args.Source, args.Format)
+	failOnError(err, "message-reader", os.Exit)
+	source = NewTransformingMessageSource(source,
 		FireHoseTransformer,
 		NewPropertiesTransformer(args.Properties))
 
-	failOnError(err, "message-reader", os.Exit)
 	err = cmdPublish(ctx, CmdPublishArg{
-		amqpURL:      args.AMQPURL,
-		exchange:     args.PubExchange,
-		routingKey:   args.PubRoutingKey,
-		headers:      args.Args,
-		fixedDelay:   args.Delay,
-		speed:        args.Speed,
-		tlsConfig:    getTLSConfig(args.InsecureTLS, args.TLSCertFile, args.TLSKeyFile, args.TLSCaFile),
-		mandatory:    args.Mandatory,
-		confirms:     args.Confirms,
-		providerFunc: provider})
+		amqpURL:    args.AMQPURL,
+		exchange:   args.PubExchange,
+		routingKey: args.PubRoutingKey,
+		headers:    args.Args,
+		fixedDelay: args.Delay,
+		speed:      args.Speed,
+		tlsConfig:  getTLSConfig(args.InsecureTLS, args.TLSCertFile, args.TLSKeyFile, args.TLSCaFile),
+		mandatory:  args.Mandatory,
+		confirms:   args.Confirms,
+		source:     source})
 	failOnError(err, "publish", os.Exit)
 }
 
 func startCmdSubscribe(ctx context.Context, args CommandLineArgs) {
-	opts := MessageReceiveFuncOptions{
+	opts := MessageSinkOptions{
 		out:              NewColorableWriter(os.Stdout),
 		format:           args.Format,
 		silent:           args.Silent,
 		optSaveDir:       args.SaveDir,
 		filenameProvider: defaultFilenameProvider,
 	}
-	messageReceiveFunc, err := createMessageReceiveFunc(opts)
+	messageSink, err := NewMessageSink(opts)
 	failOnError(err, "options", os.Exit)
 
 	termPred, err := NewLoopCountPred(args.Limit)
@@ -167,29 +167,29 @@ func startCmdSubscribe(ctx context.Context, args CommandLineArgs) {
 	failOnError(err, fmt.Sprintf("invalid message filter predicate '%s'", args.Filter), os.Exit)
 
 	err = cmdSubscribe(ctx, CmdSubscribeArg{
-		amqpURL:            args.AMQPURL,
-		queue:              args.QueueName,
-		requeue:            args.Requeue,
-		reject:             args.Reject,
-		tlsConfig:          getTLSConfig(args.InsecureTLS, args.TLSCertFile, args.TLSKeyFile, args.TLSCaFile),
-		messageReceiveFunc: messageReceiveFunc,
-		filterPred:         filterPred,
-		termPred:           termPred,
-		args:               args.Args,
-		timeout:            args.IdleTimeout,
+		amqpURL:     args.AMQPURL,
+		queue:       args.QueueName,
+		requeue:     args.Requeue,
+		reject:      args.Reject,
+		tlsConfig:   getTLSConfig(args.InsecureTLS, args.TLSCertFile, args.TLSKeyFile, args.TLSCaFile),
+		messageSink: messageSink,
+		filterPred:  filterPred,
+		termPred:    termPred,
+		args:        args.Args,
+		timeout:     args.IdleTimeout,
 	})
 	failOnError(err, "error subscribing messages", os.Exit)
 }
 
 func startCmdTap(ctx context.Context, args CommandLineArgs) {
-	opts := MessageReceiveFuncOptions{
+	opts := MessageSinkOptions{
 		out:              NewColorableWriter(os.Stdout),
 		format:           args.Format,
 		silent:           args.Silent,
 		optSaveDir:       args.SaveDir,
 		filenameProvider: defaultFilenameProvider,
 	}
-	messageReceiveFunc, err := createMessageReceiveFunc(opts)
+	messageSink, err := NewMessageSink(opts)
 	failOnError(err, "options", os.Exit)
 	termPred, err := NewLoopCountPred(args.Limit)
 	failOnError(err, "invalid message limit predicate", os.Exit)
@@ -198,12 +198,12 @@ func startCmdTap(ctx context.Context, args CommandLineArgs) {
 
 	cmdTap(ctx,
 		CmdTapArg{
-			tapConfig:          args.TapConfig,
-			tlsConfig:          getTLSConfig(args.InsecureTLS, args.TLSCertFile, args.TLSKeyFile, args.TLSCaFile),
-			messageReceiveFunc: messageReceiveFunc,
-			filterPred:         filterPred,
-			termPred:           termPred,
-			timeout:            args.IdleTimeout,
+			tapConfig:   args.TapConfig,
+			tlsConfig:   getTLSConfig(args.InsecureTLS, args.TLSCertFile, args.TLSKeyFile, args.TLSCaFile),
+			messageSink: messageSink,
+			filterPred:  filterPred,
+			termPred:    termPred,
+			timeout:     args.IdleTimeout,
 		})
 }